+++ SearchThing/src/Extensions/Filter/AvatarTagPanelPage.cs
@@ -1,0 +1,13 @@
+﻿using SearchThing.Extensions.Abstract;
+using SearchThing.Search;
+
+namespace SearchThing.Extensions.Filter;
+
+public class AvatarTagPanelPage : FilterSearchPanelPage
+{
+    public override string Tag => "Avatar";
+    protected override bool Filter(SearchableCrate searchableCrate)
+    {
+        return searchableCrate.Type == CrateType.Avatar;
+    }
+}