+++ SearchThing/staging/CHANGELOG.md
@@ -1,8 +1,16 @@
->**Version 0.1.2**
+> >>**Version 0.2.1**
+>>1. Fixed a performance issue caused by thread starvation if a lot of searches matched at the same time
+
+> >>**Version 0.2.0**
+>>1. Fixed a crash that happened when the query was empty, causing a non-unity thread to modify unity. (Reported by Alfi)
+>>2. Improved the search algorithm a bit, should now handle spaces better
+>>3. Added a spawn history
+
+> >>**Version 0.1.2**
 >>1. Fixed a bug that made the scroll buttons disappear if you closed and opened the menu
 
-> >**Version 0.1.1**
+> >>**Version 0.1.1**
 >>1. Imagine uploading the wrong build lmao
 
-> >**Version 0.1.0**
+> >>**Version 0.1.0**
 >>1. Initial Release